fixed database writing issue
