+++ .env
@@ -1,1 +1,2 @@
-PG_URI = postgres://yiwnulzh:Ti7u3rChX3E_UdmE_-syThptYEWnLQuz@suleiman.db.elephantsql.com/yiwnulzh
+# PG_URI = postgres://yiwnulzh:Ti7u3rChX3E_UdmE_-syThptYEWnLQuz@suleiman.db.elephantsql.com/yiwnulzh
+PG_URI = postgres://broybpvo:xHvnFMvX8RmNPGoK77_xllC2HcePa6TJ@suleiman.db.elephantsql.com/broybpvo
