+++ src/app.jsx
@@ -4,7 +4,8 @@ import CodeMirror from '@uiw/react-codemirror';
 // import 'codemirror/theme/dracula.css';
 import { javascript } from '@codemirror/lang-javascript';
 import { darcula } from '@uiw/codemirror-theme-darcula';
-import logo from './files/logo.png'
+
+import logo from './/files/logo.png'
 
 import { Box, Button, AppBar, Typography, Toolbar } from '@mui/material';
 import ResultsBox from './components/Results.jsx';
@@ -17,6 +18,28 @@ import { Bar, Doughnut, Line } from 'react-chartjs-2'
 const App = () => {
   const [value, setValue] = React.useState("console.log('hello world!');");
   const [results, setResults] = useState({});
+
+  const [solutions, setSolutions] = useState([])
+
+  useEffect(() => {
+    if(solutions.length) {
+      console.log("Solutions for", solutions[0].label)
+      console.log(solutions)
+    }
+  }, [solutions])
+
+  useEffect(() => {
+    async function getSolutions() {
+      const res = await fetch('/api/data')
+      const data = await res.json();
+      setSolutions(data)
+    }
+    getSolutions()
+
+  }, [])
+
+  
+
  
   const onChange = React.useCallback((val, viewUpdate) => {
     setValue(val);
@@ -60,8 +83,8 @@ const App = () => {
     </div>
 
     <div className="app-right">
-      <Solutions />
-      <SolutionsChart />
+      <Solutions solutions={solutions}/>
+      <SolutionsChart solutions={solutions}/>
     </div>
   </div>
   )
