+++ src/components/SolutionsChart.jsx
@@ -3,32 +3,13 @@ import React, { useState, useEffect } from 'react';
 import { Chart as ChartJS } from 'chart.js/auto'
 import { Bar, Doughnut, Line } from 'react-chartjs-2'
 
-const SolutionsChart = () => {
-  const [solutions, setSolutions] = useState([])
-
-  useEffect(() => {
-    if(solutions.length) {
-      console.log("Solutions for", solutions[0].label)
-      console.log(solutions)
-    }
-  }, [solutions])
-
-  useEffect(() => {
-    async function getSolutions() {
-      const res = await fetch('/api/data')
-      const data = await res.json();
-      setSolutions(data)
-    }
-    getSolutions()
-
-  }, [])
+const SolutionsChart = ({ solutions }) => {
 
   const containerStyle = {
     background: 'white', // Set the background to white
     borderRadius: '8px',
     padding: '16px',
   };
-  
 
   return (
     <div style={containerStyle}>
